requirements.txt
# Python dependencies for testing API
openai>=1.0.0
python-dotenv>=1.0.0
